pyproject.toml
[project]
name = "ai-customer-support-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.2",
    "fastapi-websocket-pubsub>=1.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.7",
    "passlib[bcrypt]>=1.7.4",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.36.0",
    "websockets>=15.0.1",
]
